sigetsop-api/requirements.txt
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
astor==0.8.1
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
bce-python-sdk==0.9.49
beautifulsoup4==4.14.2
black==25.11.0
blank==0.1.0
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorlog==6.10.1
constantly==23.10.4
cryptography==46.0.3
cssselect==1.3.0
cssselect2==0.8.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.0
daphne==4.2.1
decorator==5.2.1
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
filelock==3.19.1
fire==0.7.1
Flask==3.1.2
flask-babel==4.0.0
fonttools==4.60.1
fsspec==2025.9.0
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.1
hyperlink==21.0.0
idna==3.11
imagesize==1.4.1
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
MarkupSafe==2.1.5
matplotlib<3.9
modelscope==1.31.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
munkres==1.1.4
mypy_extensions==1.1.0
networkx==3.3
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddleocr==2.6.1.3
paddlepaddle>=2.6.1
pathspec==0.12.1
pdf2docx==0.5.8
pdf2image==1.17.0
pillow==12.0.0
pip==25.2
platformdirs==4.5.0
polars==1.35.1
polars-runtime-32==1.35.1
premailer==3.10.0
prettytable==3.16.0
protobuf==6.33.0
psutil==7.1.3
psycopg2-binary==2.9.11
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
pydyf==0.11.0
PyJWT==2.10.1
PyMuPDF==1.20.2
pyOpenSSL==25.3.0
pyparsing==3.2.5
pypdfium2==5.0.0
pyphen==0.17.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.2
pyzbar==0.1.9
qrdet==2.5
qreader==3.16
quadrilateral-fitter==1.12
RapidFuzz==3.14.3
rarfile==4.2
redis==7.1.0
requests==2.32.5
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.15.3
service-identity==24.2.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
sympy==1.14.0
termcolor==3.2.0
tifffile==2025.5.10
tinycss2==1.5.1
tinyhtml5==2.0.0
tomli==2.3.0
torch
torchvision
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typer-slim==0.20.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
ujson==5.11.0
ultralytics
ultralytics-thop
unicodedata2==17.0.0
urllib3==2.5.0
visualdl==2.5.3
wcwidth==0.2.14
weasyprint==67.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
zope.interface==8.0.1
zopfli==0.4.0
